requirements.txt
flask==2.3.3
flask-login==0.6.2
python-dotenv==1.0.0
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==1.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3 